LD n, nn where n is a 16bit reg

+++ src/core/emu/Memory.h
@@ -1,7 +1,7 @@
 #pragma once
 
 using Byte  = uint8_t;
-using Word  = uint32_t;
+using Word  = uint16_t;
 using Dword = uint32_t;
 
 #define LEN_BYTE 8u
